fixed big VIC-170 and cleaned up code
